cache needs to update, move to another process
